refactor : react-query Provider 설정 변경
