+++ src/components/app/queryClient.ts
@@ -1,0 +1,23 @@
+import { QueryClient, isServer } from "@tanstack/react-query"
+
+function makeQueryClient() {
+  return new QueryClient({
+    defaultOptions: {
+      queries: {
+        staleTime: 60 * 1000,
+      },
+    },
+  })
+}
+
+let browserQueryClient: QueryClient | undefined
+
+const getQueryClient = () => {
+  if (isServer) {
+    return makeQueryClient()
+  }
+  if (!browserQueryClient) browserQueryClient = makeQueryClient()
+  return browserQueryClient
+}
+
+export default getQueryClient
